refactor(oauth): remove oauthdb token functions
