+++ packages/fxa-auth-server/lib/oauth/validators.js
@@ -84,4 +84,7 @@ exports.ppidSeed = authServerValidators.ppidSeed.default(0);
 exports.resourceUrl = authServerValidators.resourceUrl;
 exports.pkceCodeChallengeMethod = authServerValidators.pkceCodeChallengeMethod;
 exports.pkceCodeChallenge = authServerValidators.pkceCodeChallenge;
+exports.pkceCodeVerifier = authServerValidators.pkceCodeVerifier;
 exports.refreshToken = authServerValidators.refreshToken;
+exports.authorizationCode = authServerValidators.authorizationCode;
+exports.url = authServerValidators.url;
