+++ packages/fxa-auth-server/test/local/routes/oauth.js
@@ -17,33 +17,23 @@ const JWTIdToken = require(`${ROOT_DIR}/lib/oauth/jwt_id_token`);
 const { OAUTH_SCOPE_OLD_SYNC } = require(`${ROOT_DIR}/lib/constants`);
 const MOCK_CLIENT_ID = '0123456789abcdef';
 const MOCK_USER_ID = '0123456789abcdef0123456789abcdef';
-const MOCK_ANON_ID = 'ANONYMOUSID0987654321';
 const MOCK_SCOPES = `profile https://identity.mozilla.com/apps/scoped-example ${OAUTH_SCOPE_OLD_SYNC}`;
 const MOCK_UNIX_TIMESTAMP = Math.round(Date.now() / 1000);
 const MOCK_TOKEN =
   '00112233445566778899aabbccddeeff00112233445566778899aabbccddeeff';
 const MOCK_JWT =
   '001122334455.66778899aabbccddeeff00112233445566778899.aabbccddeeff';
-const MOCK_AUTHORIZATION_CODE =
-  'aaaaaabbbbbbccccccddddddeeeeeeffaaaaaabbbbbbccccccddddddeeeeeeff';
-const MOCK_TOKEN_RESPONSE = {
-  access_token: MOCK_TOKEN,
-  scope: MOCK_SCOPES,
-  token_type: 'bearer',
-  expires_in: 500,
-  auth_at: 123456,
-};
 
 describe('/oauth/ routes', () => {
   let mockOAuthDB, mockDB, mockLog, mockConfig, sessionToken;
 
   async function loadAndCallRoute(path, request) {
     const routes = require('../../../lib/routes/oauth/proxied')(
-      mockLog,
       mockConfig,
-      mockOAuthDB,
-      mockDB
-    ).concat(require('../../../lib/routes/oauth')(mockLog, mockConfig));
+      mockOAuthDB
+    ).concat(
+      require('../../../lib/routes/oauth')(mockLog, mockConfig, mockDB, {}, {})
+    );
     const route = await getRoute(routes, path);
     if (route.config.validate.payload) {
       // eslint-disable-next-line require-atomic-updates
@@ -223,7 +213,6 @@ describe('/oauth/ routes', () => {
 
   describe('/oauth/authorization', () => {
     it('can refuse to authorize new grants for selected OAuth clients', async () => {
-      mockOAuthDB = mocks.mockOAuthDB();
       mockConfig.oauth.disableNewConnectionsForClients = [MOCK_CLIENT_ID];
       sessionToken = await mockSessionToken();
       const mockRequest = mocks.mockRequest({
@@ -241,498 +230,6 @@ describe('/oauth/ routes', () => {
         assert.equal(err.output.statusCode, 503);
         assert.equal(err.errno, error.ERRNO.DISABLED_CLIENT_ID);
       }
-    });
-  });
-
-  describe('/oauth/token', () => {
-    it('calls oauthdb.grantTokensFromSessionToken when told nothing else about the grant type', async () => {
-      mockOAuthDB = mocks.mockOAuthDB({
-        grantTokensFromSessionToken: sinon.spy(async () => {
-          return MOCK_TOKEN_RESPONSE;
-        }),
-      });
-      mockDB = mocks.mockDB({
-        ecosystemAnonId: MOCK_ANON_ID,
-        email: 'foo@example.com',
-        uid: MOCK_USER_ID,
-      });
-      sessionToken = await mockSessionToken();
-      const mockRequest = mocks.mockRequest({
-        credentials: sessionToken,
-        payload: {
-          client_id: MOCK_CLIENT_ID,
-          scope: MOCK_SCOPES,
-        },
-      });
-      const mockEmitMetricsEvent = sinon.stub(mockRequest, 'emitMetricsEvent');
-      const resp = await loadAndCallRoute('/oauth/token', mockRequest);
-      assert.calledOnce(mockOAuthDB.grantTokensFromSessionToken);
-      assert.calledWithExactly(
-        mockOAuthDB.grantTokensFromSessionToken,
-        sessionToken,
-        {
-          client_id: MOCK_CLIENT_ID,
-          scope: MOCK_SCOPES,
-          grant_type: 'fxa-credentials',
-          access_type: 'online',
-        }
-      );
-      assert.called(mockDB.account);
-      assert.calledWithExactly(mockDB.account, MOCK_USER_ID);
-      assert.calledOnce(mockEmitMetricsEvent);
-      assert.calledWithMatch(
-        mockEmitMetricsEvent.firstCall,
-        'oauth.token.created',
-        {
-          grantType: 'fxa-credentials',
-          uid: MOCK_USER_ID,
-          ecosystemAnonId: MOCK_ANON_ID,
-          service: MOCK_CLIENT_ID,
-        }
-      );
-      assert.deepEqual(resp, MOCK_TOKEN_RESPONSE);
-    });
-
-    it('calls oauthdb.grantTokensFromAuthorizationCode when `code` input is provided', async () => {
-      mockOAuthDB = mocks.mockOAuthDB({
-        grantTokensFromAuthorizationCode: sinon.spy(async () => {
-          return MOCK_TOKEN_RESPONSE;
-        }),
-      });
-      mockDB = mocks.mockDB({
-        ecosystemAnonId: MOCK_ANON_ID,
-        email: 'foo@example.com',
-        uid: MOCK_USER_ID,
-      });
-      const mockRequest = mocks.mockRequest({
-        credentials: sessionToken,
-        payload: {
-          client_id: MOCK_CLIENT_ID,
-          client_secret: 'ABCDEF',
-          code: MOCK_AUTHORIZATION_CODE,
-        },
-      });
-      const mockEmitMetricsEvent = sinon.stub(mockRequest, 'emitMetricsEvent');
-      const resp = await loadAndCallRoute('/oauth/token', mockRequest);
-      assert.calledOnce(mockOAuthDB.grantTokensFromAuthorizationCode);
-      assert.calledWithExactly(mockOAuthDB.grantTokensFromAuthorizationCode, {
-        client_id: MOCK_CLIENT_ID,
-        client_secret: 'ABCDEF',
-        code: MOCK_AUTHORIZATION_CODE,
-        grant_type: 'authorization_code',
-      });
-      assert.called(mockDB.account);
-      assert.calledWithExactly(mockDB.account, MOCK_USER_ID);
-      assert.calledOnce(mockEmitMetricsEvent);
-      assert.calledWithMatch(
-        mockEmitMetricsEvent.firstCall,
-        'oauth.token.created',
-        {
-          grantType: 'authorization_code',
-          uid: MOCK_USER_ID,
-          ecosystemAnonId: MOCK_ANON_ID,
-          service: MOCK_CLIENT_ID,
-        }
-      );
-      assert.deepEqual(resp, MOCK_TOKEN_RESPONSE);
-    });
-
-    it('calls oauthdb.grantTokensFromAuthorizationCode when `grant_type=authorization_code`', async () => {
-      mockOAuthDB = mocks.mockOAuthDB({
-        grantTokensFromAuthorizationCode: sinon.spy(async () => {
-          return MOCK_TOKEN_RESPONSE;
-        }),
-      });
-      mockDB = mocks.mockDB({
-        ecosystemAnonId: MOCK_ANON_ID,
-        email: 'foo@example.com',
-        uid: MOCK_USER_ID,
-      });
-      const mockRequest = mocks.mockRequest({
-        credentials: sessionToken,
-        payload: {
-          client_id: MOCK_CLIENT_ID,
-          client_secret: 'ABCDEF',
-          code: MOCK_AUTHORIZATION_CODE,
-          grant_type: 'authorization_code',
-        },
-      });
-      const mockEmitMetricsEvent = sinon.stub(mockRequest, 'emitMetricsEvent');
-      const resp = await loadAndCallRoute('/oauth/token', mockRequest);
-      assert.calledOnce(mockOAuthDB.grantTokensFromAuthorizationCode);
-      assert.calledWithExactly(mockOAuthDB.grantTokensFromAuthorizationCode, {
-        client_id: MOCK_CLIENT_ID,
-        client_secret: 'ABCDEF',
-        code: MOCK_AUTHORIZATION_CODE,
-        grant_type: 'authorization_code',
-      });
-
-      assert.called(mockDB.account);
-      assert.calledWithExactly(mockDB.account, MOCK_USER_ID);
-      assert.calledOnce(mockEmitMetricsEvent);
-      assert.calledWithMatch(
-        mockEmitMetricsEvent.firstCall,
-        'oauth.token.created',
-        {
-          grantType: 'authorization_code',
-          uid: MOCK_USER_ID,
-          ecosystemAnonId: MOCK_ANON_ID,
-          service: MOCK_CLIENT_ID,
-        }
-      );
-
-      assert.deepEqual(resp, MOCK_TOKEN_RESPONSE);
-    });
-
-    it('calls oauthdb.grantTokensFromRefreshToken when `grant_type=refresh_token`', async () => {
-      mockOAuthDB = mocks.mockOAuthDB({
-        grantTokensFromRefreshToken: sinon.spy(async () => {
-          return MOCK_TOKEN_RESPONSE;
-        }),
-      });
-      mockDB = mocks.mockDB({
-        ecosystemAnonId: MOCK_ANON_ID,
-        email: 'foo@example.com',
-        uid: MOCK_USER_ID,
-      });
-      const mockRequest = mocks.mockRequest({
-        credentials: sessionToken,
-        payload: {
-          client_id: MOCK_CLIENT_ID,
-          client_secret: 'ABCDEF',
-          refresh_token: MOCK_AUTHORIZATION_CODE,
-          grant_type: 'refresh_token',
-        },
-      });
-      const mockEmitMetricsEvent = sinon.stub(mockRequest, 'emitMetricsEvent');
-      const resp = await loadAndCallRoute('/oauth/token', mockRequest);
-      assert.calledOnce(mockOAuthDB.grantTokensFromRefreshToken);
-      assert.calledWithExactly(mockOAuthDB.grantTokensFromRefreshToken, {
-        client_id: MOCK_CLIENT_ID,
-        client_secret: 'ABCDEF',
-        refresh_token: MOCK_AUTHORIZATION_CODE,
-        grant_type: 'refresh_token',
-      });
-      assert.called(mockDB.account);
-      assert.calledWithExactly(mockDB.account, MOCK_USER_ID);
-      assert.calledOnce(mockEmitMetricsEvent);
-      assert.calledWithMatch(
-        mockEmitMetricsEvent.firstCall,
-        'oauth.token.created',
-        {
-          grantType: 'refresh_token',
-          uid: MOCK_USER_ID,
-          ecosystemAnonId: MOCK_ANON_ID,
-          service: MOCK_CLIENT_ID,
-        }
-      );
-      assert.deepEqual(resp, MOCK_TOKEN_RESPONSE);
-    });
-
-    it('calls oauthdb.grantTokensFromSessionToken when `grant_type=fxa-credentials`', async () => {
-      mockOAuthDB = mocks.mockOAuthDB({
-        grantTokensFromSessionToken: sinon.spy(async () => {
-          return MOCK_TOKEN_RESPONSE;
-        }),
-      });
-      mockDB = mocks.mockDB({
-        ecosystemAnonId: MOCK_ANON_ID,
-        email: 'foo@example.com',
-        uid: MOCK_USER_ID,
-      });
-      sessionToken = await mockSessionToken();
-      const mockRequest = mocks.mockRequest({
-        credentials: sessionToken,
-        payload: {
-          client_id: MOCK_CLIENT_ID,
-          scope: MOCK_SCOPES,
-          grant_type: 'fxa-credentials',
-        },
-      });
-      const mockEmitMetricsEvent = sinon.stub(mockRequest, 'emitMetricsEvent');
-      const resp = await loadAndCallRoute('/oauth/token', mockRequest);
-      assert.calledOnce(mockOAuthDB.grantTokensFromSessionToken);
-      assert.calledWithExactly(
-        mockOAuthDB.grantTokensFromSessionToken,
-        sessionToken,
-        {
-          client_id: MOCK_CLIENT_ID,
-          scope: MOCK_SCOPES,
-          grant_type: 'fxa-credentials',
-          access_type: 'online',
-        }
-      );
-      assert.called(mockDB.account);
-      assert.calledWithExactly(mockDB.account, MOCK_USER_ID);
-      assert.calledOnce(mockEmitMetricsEvent);
-      assert.calledWithMatch(
-        mockEmitMetricsEvent.firstCall,
-        'oauth.token.created',
-        {
-          grantType: 'fxa-credentials',
-          uid: MOCK_USER_ID,
-          ecosystemAnonId: MOCK_ANON_ID,
-          service: MOCK_CLIENT_ID,
-        }
-      );
-      assert.deepEqual(resp, MOCK_TOKEN_RESPONSE);
-    });
-
-    it('refuses to call oauthdb.grantTokensFromSessionToken without a valid sessionToken', async () => {
-      mockOAuthDB = mocks.mockOAuthDB({
-        grantTokensFromSessionToken: sinon.spy(async () => {
-          return MOCK_TOKEN_RESPONSE;
-        }),
-      });
-      const mockRequest = mocks.mockRequest({
-        credentials: null,
-        payload: {
-          client_id: MOCK_CLIENT_ID,
-          scope: MOCK_SCOPES,
-          grant_type: 'fxa-credentials',
-        },
-      });
-      const mockEmitMetricsEvent = sinon.stub(mockRequest, 'emitMetricsEvent');
-      try {
-        await loadAndCallRoute('/oauth/token', mockRequest);
-        throw new Error('should have thrown');
-      } catch (err) {
-        assert.equal(err.errno, error.ERRNO.INVALID_TOKEN);
-      }
-      assert.equal(mockEmitMetricsEvent.callCount, 0);
-      assert.equal(mockOAuthDB.grantTokensFromSessionToken.callCount, 0);
-    });
-
-    it('sets the service event property to "sync" when the client is an old sync client', async () => {
-      mockConfig.oauth.oldSyncClientIds = [MOCK_CLIENT_ID];
-      mockOAuthDB = mocks.mockOAuthDB({
-        grantTokensFromSessionToken: sinon.spy(async () => {
-          return { ...MOCK_TOKEN_RESPONSE, scope: OAUTH_SCOPE_OLD_SYNC };
-        }),
-      });
-      mockDB = mocks.mockDB({
-        ecosystemAnonId: MOCK_ANON_ID,
-        email: 'foo@example.com',
-        uid: MOCK_USER_ID,
-      });
-      sessionToken = await mockSessionToken();
-      const mockRequest = mocks.mockRequest({
-        credentials: sessionToken,
-        payload: {
-          client_id: MOCK_CLIENT_ID,
-          scope: OAUTH_SCOPE_OLD_SYNC,
-          grant_type: 'fxa-credentials',
-        },
-      });
-      const mockEmitMetricsEvent = sinon.stub(mockRequest, 'emitMetricsEvent');
-      await loadAndCallRoute('/oauth/token', mockRequest);
-      assert.calledWithMatch(
-        mockEmitMetricsEvent.secondCall,
-        'account.signed',
-        {
-          uid: MOCK_USER_ID,
-          device_id: undefined,
-          service: 'sync',
-        }
-      );
-    });
-
-    describe('"flow.complete" event', async () => {
-      let metricsContext, mockLog;
-
-      beforeEach(async () => {
-        mockDB = mocks.mockDB({
-          ecosystemAnonId: MOCK_ANON_ID,
-          email: 'foo@example.com',
-          uid: MOCK_USER_ID,
-        });
-        sessionToken = await mockSessionToken();
-        metricsContext = mocks.mockMetricsContext({
-          gather: sinon.spy(() =>
-            Promise.resolve({
-              device_id: 'foo',
-              flow_id: 'bar',
-              flowBeginTime: 123,
-              flowCompleteSignal: 'account.signed',
-            })
-          ),
-        });
-        mockLog = mocks.mockLog();
-      });
-
-      describe('desktop sync using "fxa-credentials" grant', async () => {
-        it('emits the event when the OLD_SYNC scope alone is requested', async () => {
-          const mockTokenResponseOldSyncScope = {
-            ...MOCK_TOKEN_RESPONSE,
-            scope: OAUTH_SCOPE_OLD_SYNC,
-          };
-          mockConfig.oauth.oldSyncClientIds = [MOCK_CLIENT_ID];
-          mockOAuthDB = mocks.mockOAuthDB({
-            grantTokensFromSessionToken: sinon.spy(async () => {
-              return mockTokenResponseOldSyncScope;
-            }),
-          });
-          const mockRequest = mocks.mockRequest({
-            log: mockLog,
-            credentials: sessionToken,
-            payload: {
-              client_id: MOCK_CLIENT_ID,
-              scope: OAUTH_SCOPE_OLD_SYNC,
-            },
-            metricsContext,
-          });
-          const resp = await loadAndCallRoute('/oauth/token', mockRequest);
-          assert.isTrue(
-            mockLog.flowEvent.args.some(
-              (args) => args[0].event && args[0].event === 'flow.complete'
-            )
-          );
-          assert.deepEqual(resp, mockTokenResponseOldSyncScope);
-        });
-
-        it('does not emit the event when OLD_SYNC is not in the requested scopes', async () => {
-          mockOAuthDB = mocks.mockOAuthDB({
-            grantTokensFromSessionToken: sinon.spy(async () => {
-              return { ...MOCK_TOKEN_RESPONSE, scope: 'somethingelse' };
-            }),
-          });
-          const mockRequest = mocks.mockRequest({
-            log: mockLog,
-            credentials: sessionToken,
-            payload: {
-              client_id: MOCK_CLIENT_ID,
-              scope: MOCK_SCOPES,
-            },
-            metricsContext,
-          });
-          await loadAndCallRoute('/oauth/token', mockRequest);
-          assert.isFalse(
-            mockLog.flowEvent.args.some(
-              (args) => args[0].event && args[0].event === 'flow.complete'
-            )
-          );
-        });
-
-        it('does not emit the event when the both OLD_SYNC and `profile` scopes are requested', async () => {
-          mockOAuthDB = mocks.mockOAuthDB({
-            grantTokensFromSessionToken: sinon.spy(async () => {
-              return MOCK_TOKEN_RESPONSE;
-            }),
-          });
-          const mockRequest = mocks.mockRequest({
-            log: mockLog,
-            credentials: sessionToken,
-            payload: {
-              client_id: MOCK_CLIENT_ID,
-              scope: MOCK_SCOPES,
-            },
-            metricsContext,
-          });
-          await loadAndCallRoute('/oauth/token', mockRequest);
-          assert.isFalse(
-            mockLog.flowEvent.args.some(
-              (args) => args[0].event && args[0].event === 'flow.complete'
-            )
-          );
-        });
-      });
-
-      describe('mobile sync', async () => {
-        it('does not emit the event for the authorization code grant', async () => {
-          mockOAuthDB = mocks.mockOAuthDB({
-            grantTokensFromAuthorizationCode: sinon.spy(async () => {
-              return MOCK_TOKEN_RESPONSE;
-            }),
-          });
-          mockDB = mocks.mockDB({
-            ecosystemAnonId: MOCK_ANON_ID,
-            email: 'foo@example.com',
-            uid: MOCK_USER_ID,
-          });
-          const mockRequest = mocks.mockRequest({
-            log: mockLog,
-            metricsContext,
-            credentials: sessionToken,
-            payload: {
-              client_id: MOCK_CLIENT_ID,
-              code: MOCK_AUTHORIZATION_CODE,
-              code_verifier:
-                'LTvYq_2MBc2XAX_unyHRMST4x7XWSALLbeXS8JUNzPrEoXYsD1oyyDjErw',
-              grant_type: 'authorization_code',
-            },
-          });
-          await loadAndCallRoute('/oauth/token', mockRequest);
-          assert.isFalse(
-            mockLog.flowEvent.args.some(
-              (args) => args[0].event && args[0].event === 'flow.complete'
-            )
-          );
-        });
-
-        it('does not emit the event for the refresh_token grant requesting profile scope', async () => {
-          mockOAuthDB = mocks.mockOAuthDB({
-            grantTokensFromRefreshToken: sinon.spy(async () => {
-              return { ...MOCK_TOKEN_RESPONSE, scope: 'profile' };
-            }),
-          });
-          mockDB = mocks.mockDB({
-            ecosystemAnonId: MOCK_ANON_ID,
-            email: 'foo@example.com',
-            uid: MOCK_USER_ID,
-          });
-          const mockRequest = mocks.mockRequest({
-            log: mockLog,
-            metricsContext,
-            credentials: sessionToken,
-            payload: {
-              client_id: MOCK_CLIENT_ID,
-              refresh_token: MOCK_AUTHORIZATION_CODE,
-              grant_type: 'refresh_token',
-              scope: 'profile',
-            },
-          });
-          await loadAndCallRoute('/oauth/token', mockRequest);
-          assert.isFalse(
-            mockLog.flowEvent.args.some(
-              (args) => args[0].event && args[0].event === 'flow.complete'
-            )
-          );
-        });
-
-        it('emits the event for the refresh_token grant requesting OLD_SYNC scope', async () => {
-          mockConfig.oauth.oldSyncClientIds = [MOCK_CLIENT_ID];
-          mockOAuthDB = mocks.mockOAuthDB({
-            grantTokensFromRefreshToken: sinon.spy(async () => {
-              return { ...MOCK_TOKEN_RESPONSE, scope: OAUTH_SCOPE_OLD_SYNC };
-            }),
-          });
-          const mockRequest = mocks.mockRequest({
-            metricsContext,
-            credentials: sessionToken,
-            payload: {
-              client_id: MOCK_CLIENT_ID,
-              refresh_token: MOCK_AUTHORIZATION_CODE,
-              grant_type: 'refresh_token',
-              scope: OAUTH_SCOPE_OLD_SYNC,
-            },
-          });
-          const mockEmitMetricsEvent = sinon.stub(
-            mockRequest,
-            'emitMetricsEvent'
-          );
-          await loadAndCallRoute('/oauth/token', mockRequest);
-          assert.calledTwice(mockEmitMetricsEvent);
-          assert.calledWithMatch(
-            mockEmitMetricsEvent.secondCall,
-            'account.signed',
-            {
-              uid: MOCK_USER_ID,
-              device_id: undefined,
-              service: 'sync',
-            }
-          );
-        });
-      });
     });
   });
